deployment/requirements.txt
streamlit == 1.38.0
pandas == 2.2.2
seaborn == 0.13.2
matplotlib == 3.9.2
numpy == 2.0.1
scikit-learn==1.5.2
Pillow 
plotly
imblearn